fix: 주문 가격 반영 총금액  오류 수정
